backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
aiosqlite==0.19.0
httpx==0.26.0
alembic==1.13.1
slowapi==0.1.9
python-dotenv==1.0.0
python-frontmatter==1.0.1
